refactored folder structure2
