+++ apps/web/pages/rooms/[roomId].tsx
@@ -1,8 +1,5 @@
 import dynamic from 'next/dynamic';
 import { useState } from 'react';
-import { useRecoilState, useRecoilValue } from 'recoil';
-import { canvasState, polaroidState } from 'web/atoms';
-
 import { ActionBar } from 'web/components/layout/ActionBar';
 import { MenuBar } from 'web/components/layout/MenuBar';
 import { EditModal } from 'web/components/layout/EditModal';
@@ -16,8 +13,6 @@ interface Props {}
 
 function WhiteBoard({}: Props) {
   const [isEditOpen, setEditOpen] = useState<boolean>(false);
-  const [canvas, setCanvas] = useRecoilState(canvasState);
-  const polaroid = useRecoilValue(polaroidState);
 
   return (
     <>
