+++ apps/web/shared/hooks/index.ts
@@ -1,0 +1,2 @@
+export { useCanvasDragEvent } from './useCanvasDragEvent';
+export { useCanvasDblClickEvent } from './useCanvasDblClickEvent';
