+++ apps/web/shared/hooks/useCanvasDblClickEvent.ts
@@ -1,0 +1,1 @@
+export function useCanvasDblClickEvent() {}
